feat-1.4.0/AB#42918 Create new file for custom notification scheduler
